Create RESTful API (GET)
API that returns user data when a GET request is made to /users/<id>. If the user ID exists in our data, it returns the user data. Otherwise, it returns an error message with a 404 status code.
Remember to replace the sample data with your actual data and add more functionalities as per your requirements. For example, you might want to add endpoints for creating, updating, and deleting users.