Create **MODULE
This script includes all the aspects we discussed:

Secure Password Input: We’re using the getpass module to securely get the password from the user. This prevents the password from being displayed on the screen.

Secure Salt Generation: We’re using the os.urandom function to generate a secure random salt.

Advanced Error Handling: We’re passing exc_info=True to logging.error to include the traceback in the log if an error occurs.

Advanced Logging Techniques: We’re using a more detailed logging format that includes the timestamp, the severity level of the log message, and the log message itself.

Secure Routing: We’re passing verify=True to requests.get to verify the SSL certificate of the server we’re making a request to. This helps ensure that we’re communicating with the server we think we’re communicating with and that the data we’re sending and receiving is encrypted.

Secure Storage of the Encryption Key: We’re using the keyring library to securely store the encryption key in the system’s keyring.